pyproject.toml
[tool.poetry]
name = "ylabproject"
version = "0.1.0"
description = ""
authors = ["Тимофей Боровец <t.m.borovets@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.19"}
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
pytest = "^7.4.0"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.1"
aioredis = "^2.0.1"
alembic = "^1.11.1"
redis = "^4.6.0"
pytest-order = "^1.1.0"
flake8 = "^6.1.0"
flake8-docstrings = "^1.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
